Consolidate ParquetService and SendGridEvent into shared project
- Moved SendgridParquetLogger/Services/ParquetService.cs to SendgridParquet.Shared/ParquetService.cs
- Moved SendgridParquetLogger/Models/SendGridEvent.cs to SendgridParquet.Shared/SendGridEvent.cs
- Added ReadRowGroupEventsAsync method to ParquetService from CompactionService
- Updated CompactionService to use shared ParquetService via dependency injection
- Removed duplicate SendgridParquetViewer/Models/SendGridEvent.cs (59 lines)
- Added missing @using directives to fix compilation errors
- Consolidated Parquet.Net dependency to shared project
- Enhanced SendGridEvent model with Pool structure and SelectColumns helper
