+++ SendgridParquet.Shared/SendgridParquet.Shared.csproj
@@ -6,9 +6,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="9.0.0" />
-    <PackageReference Include="ZLogger" Version="2.5.2" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.8" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="9.0.8" />
+    <PackageReference Include="Parquet.Net" Version="5.2.0" />
+    <PackageReference Include="ZLogger" Version="2.5.10" />
   </ItemGroup>
 
 </Project>
