+++ SendgridParquetViewer/Components/Pages/Compaction.razor
@@ -1,6 +1,7 @@
 @page "/compaction"
 @rendermode InteractiveServer
 @using SendgridParquetViewer.Models
+@using SendgridParquetViewer.Services
 @inject ILogger<Compaction> Logger
 @inject CompactionService CompactionService
