+++ SendgridParquetViewer/SendgridParquetViewer.csproj
@@ -13,8 +13,6 @@
     <PackageReference Include="Microsoft.FluentUI.AspNetCore.Components" Version="4.12.1" />
     <PackageReference Include="Microsoft.Identity.Web" Version="3.14.0" />
     <PackageReference Include="Microsoft.Identity.Web.UI" Version="3.14.0" />
-    <PackageReference Include="Parquet.Net" Version="5.2.0" />
-    <PackageReference Include="ZLogger" Version="2.5.10" />
   </ItemGroup>
 
   <ItemGroup>
